Define _topdir when calling rpmbuild.
This patch fixes a problem with running spectool from within a SCons
build. SCons cleans up the environment of the commands it runs,
including the $HOME variable.  RPM by default sets the _topdir
variable to "%(echo $HOME)/rpmbuild".  If $HOME isn't set the rpmbuild
instance that is called from spectool will error out because it tries
to create "/rpmbuild".  Explicitly setting _topdir when rpmbuild is
called fixes this problem.

Signed-off-by: Jeffrey C. Ollie <jeff@ocjtech.us>